requirements.txt
Django==2.0.1
django-choices==1.4.4
django-filter==1.1.0
djangorestframework==3.7.7
django-rest-swagger==2.1.2
mysqlclient==1.3.12
raven==6.6.0
django-redis==4.9.0
django-health-check==3.5.1


# user sms
git+https://github.com/predawning/user_sms.git#egg=userSms

